+++ packages/algob/src/builtin-tasks/init.ts
@@ -7,16 +7,21 @@ export default function (): void {
 		.addPositionalParam<string>("newProjectLocation", "Location of the new project")
 		.addFlag("typescript", "Initializes a new typescript project in the given directory")
 		.addFlag("infrastructure", "Initializes a new project without infrastructure folder")
+		.addFlag("npm", "Use npm instead for yarn")
 		.setAction(
 			async (
 				{
 					newProjectLocation,
 					typescript,
 					infrastructure,
-				}: { newProjectLocation: string; typescript: boolean; infrastructure: boolean },
+					npm,
+				}: {
+					newProjectLocation: string;
+					typescript: boolean;
+					infrastructure: boolean;
+					npm: boolean;
+				},
 				_
-			) => {
-				await createProject(newProjectLocation, typescript, infrastructure);
-			}
+			) => await createProject(newProjectLocation, typescript, infrastructure, npm)
 		);
 }