+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Features, Bug Fixes, API Breaking, Deprecated, Infrastructure, Template Updates
 - Updated yarn to v3.2.1
 - Changed default sample project license to ISC
 
+### Features
+
+- Add `--npm` flag to `algob init` and `algob unbox`. Note: by default we will use `yarn`.
+- Improved `algob/project-dev-script.sh` which is script setting up a local project.
+
 ### Bug Fixes
 
 - Fix `txn AssetSender` should return zero address by default.